Refactor YAML parsing (#107)
* Fix coarsest level solver parser
* Add new unit tests related to YAML parsing
* Refactoring MGR parsing
* Add AMG as G-relaxation in MGR
* Add CLI support to YAML config options
* Add tests and docs for CLI args mode

+++ src/stats.c
@@ -712,7 +712,7 @@ StatsPrint(int print_level)
       return;
    }
 
-   const char *scale = active_stats->use_millisec ? "[ms]" : "[s]";
+   const char *scale = ((int)active_stats->use_millisec) ? "[ms]" : "[s]";
 
    PRINT_EQUAL_LINE(MAX_DIVISOR_LENGTH)
    printf("\n\nSTATISTICS SUMMARY:\n\n");
